+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 #### Fixed
 - Endpoints for getting player 170 statistics
 - Ensure that currenet player is set correctly on tournament playoffs
+- Don't include legs that hit max rounds as highest checkout
 
 ## [2.9.0] - 2025-04-06
 #### Feature